Move each EntryState in the graph under its own Mutex (#6095)
